cloud_functions/label-latest-run/requirements.txt
functions-framework==3.*
pandas==1.3.3
numpy==1.21.4
google-cloud-bigquery==3.4.2
google-cloud-storage==2.5.0
joblib==1.1.0
db-dtypes
scikit-learn==1.0.2